ocr/requirements.txt
rapidocr_onnxruntime==1.4.4
pyzbar==0.1.9
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
Pillow>=10.0.0
